refactor: split variable loading from shared `DocumentListModal`
